+++ src/Support/Model/DataModelAnalyser.php
@@ -188,12 +188,20 @@ class DataModelAnalyser
         // In case the value object has no attributes, we must register the attribute prefix with a null value.
         $this->propertyMap['attributes'] = array_merge(
             $this->propertyMap['attributes'],
-            $self->getAttributesPropertyMap() ?: [$attributePrefix => $propertyPrefix]
+            [$attributePrefix => $propertyPrefix],
+            $self->getAttributesPropertyMap(),
         );
-        $this->resourceValueMap['attributes'] = array_merge(
-            $this->resourceValueMap['attributes'],
-            $self->getAttributeValues() ?: [$attributePrefix => null]
-        );
+        if (null === $valueObject) {
+            $this->resourceValueMap['attributes'] = array_merge(
+                $this->resourceValueMap['attributes'],
+                [$attributePrefix => null]
+            );
+        } else {
+            $this->resourceValueMap['attributes'] = array_merge(
+                $this->resourceValueMap['attributes'],
+                $self->getAttributeValues() ?: [$attributePrefix => []]
+            );
+        }
     }
 
     private function registerAttributeValue(