+++ src/Support/Model/ValueObjectFactoryTrait.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Dogado\JsonApi\Support\Model;
+
+trait ValueObjectFactoryTrait
+{
+    public static function create(): self
+    {
+        return new self();
+    }
+}