+++ tests/Converter/ModelConverterTest.php
@@ -30,9 +30,8 @@ class ModelConverterTest extends TestCase
             'values' => [
                 'number' => $model->getValueObject()->getTest(),
             ],
-            'empty-values' => [
-                'number' => null,
-            ],
+            'empty-values' => null,
+            'valueObjectWithoutAttributes' => [],
             'createdAt' => $date->format(DateTimeInterface::ATOM),
             'updatedAt' => null,
         ]);